Problem with html.find
I changed this method because after months of my app working correctly with BeautifulSoup and Genius API it stopped working. Before the method, scrapSongUrl did the work with no error but now is giving an AttributeError, lyrics always None. For now, I only did a try and except in order for the app to work .